requirements.txt
ollama>=0.4.5
python-dotenv>=1.0.1
requests>=2.32.3